requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-channels==0.7.0
django-redis==5.4.0
django-restframework==0.0.1
djangochannelsrestframework==1.2.0
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
redis==5.0.6
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==6.4.post2